% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_report.R
\name{project_report}
\alias{project_report}
\title{R4 I&M Project Report (PR) format.}
\usage{
project_report(..., keep_tex = FALSE)
}
\arguments{
\item{...}{Arguments to \code{bookdown::pdf_book}}

\item{keep_tex}{A boolean toggle to select whether intermediate
LaTeX files are to be kept, defaults to \code{FALSE}}
}
\value{
output format to pass to \code{\link[bookdown:render_book]{render_book}}
}
\description{
Format for creating a R4 Inventory and Monitoring Project Report
}
